ai_avatar_chatbot/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
websockets==12.0
pydantic==2.5.0
numpy==1.26.3
aiohttp>=3.9.0

# Text-to-Speech (TTS)
TTS>=0.22.0
pyttsx3==2.90

# LLM Integration
ollama>=0.1.0
openai>=1.3.0
langchain>=0.1.0

# RAG - Embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0

# Vector Database
chromadb>=0.4.0
lancedb>=0.3.0

# Optional: Avatar/Lip-sync
opencv-python>=4.8.0
torch>=2.1.0
torchvision>=0.16.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
